% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ab617}
\alias{ab617}
\title{AB617 Communities}
\format{
A data frame with 6 rows and 2 variables:
\describe{
\item{\code{Name}}{character Community name}
\item{\code{geometry}}{list Polygon geometry}
}
}
\usage{
ab617
}
\description{
Boundaries of six vulnerable communities identified in A.B. 617 targeted for environmental remediation and environmental justice-related support.
}
\details{
Boundaries of vulnerable communities.
}
\keyword{datasets}
