% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_grid_intersection.R
\name{find_grid_intersection}
\alias{find_grid_intersection}
\title{Find grid intersection;}
\usage{
find_grid_intersection(points, grid = aqmdR::sc_grid)
}
\arguments{
\item{points}{An sf object representing point locations.}

\item{grid}{An sf object representing the grid. Defaults to \code{sc_grid}.}
}
\value{
An sf object with a new column \code{grid_id} representing the ID of the intersecting grid cell.
}
\description{
This function finds the intersection between points and a grid, assigning a grid ID to each point.
}
