% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_to_sf.R
\name{grid_to_sf}
\alias{grid_to_sf}
\title{Grid to SF;}
\usage{
grid_to_sf(sum_emissions, pollutant)
}
\arguments{
\item{sum_emissions}{Summed emissions in a 156x102 grid, generated, for example from the function \code{sum_pt_emissions}}

\item{pollutant}{Pollutant of interest}
}
\description{
Turns a 156x102 array -- usually emissions summed from CMAQ output -- into a sf object to be used for mapping/plotting. Uses a simple loop to add in emissions data into a template (\code{aqmdR::sc_grid}), and returns the resulting object with a \code{geometry} field. Output files are typically \code{.mat} files, which require the package \code{R.matlab} to read into R.
}
