% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sc_emissions.R
\name{plot_grid}
\alias{plot_grid}
\title{Plot emissions in the South Coast Air Basin}
\usage{
plot_grid(
  sc_df,
  pollutant,
  units = "kg/day",
  sum_emission = T,
  plot_max = 0,
  save_plot = F,
  plot_name = NA
)
}
\arguments{
\item{sc_df}{sf object with of summed emissions}

\item{pollutant}{Character. Pollutant (or any field) to be plotted.}

\item{units}{Character. Unit of the emissions measurements. Default is kg/day.}

\item{sum_emission}{boolean. If TRUE, display the sum of emissions (in tons/day) as a subtitle of the plot}

\item{save_plot}{boolean. If TRUE, save a copy of the plot}

\item{plot_name}{Character. File name for the plot to save to (requires \code{save_plot} to be TRUE)}
}
\description{
Wrapper around ggplot to plot pollutions in a 4km x 4km grid of the South Coast Air Basin
}
