% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ROWPTc_pt}
\alias{ROWPTc_pt}
\title{Domain stack}
\format{
An object of class \code{ncdf4}
}
\usage{
ROWPTc_pt
}
\description{
Matrix used to
}
\details{
To be used when calculating emissions from CMAQ output. The fields \code{COLPTc_pt} and \code{ROWPTc_pt} are used after emissions have been summed together and "flattened" from a 2D matrix to a 1D vector.
}
\keyword{datasets}
