% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc_grid}
\alias{sc_grid}
\title{South Coast Grid}
\format{
A data frame with 15912 rows and 6 variables:
\describe{
\item{\code{Id}}{integer COLUMN_DESCRIPTION}
\item{\code{IY}}{integer Y coordinate}
\item{\code{IX}}{integer X coordinate}
\item{\code{Am2}}{integer Area}
\item{\code{IJ}}{integer Point number}
\item{\code{geometry}}{list Polygon geometry}
}
}
\usage{
sc_grid
}
\description{
4 km by 4 km grid of the South Coast Air Basin bounding box.
}
\details{
156x102 grid, matching the output from CMAQ. Does not contain any air quality data, but is to be used as a "mask" to append geospatial information to CMAQ output.
}
\keyword{datasets}
