% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{south_coast}
\alias{south_coast}
\title{South Coast Air Basin}
\format{
A data frame with 1 rows and 1 variables:
\describe{
\item{\code{geometry}}{list Polygon geometry}
}
}
\usage{
south_coast
}
\description{
sf object of the South Coast Air Basin
}
\details{
Single polygon of South Coast Air Basin boundaries. Excludes Catalina Island.
}
\keyword{datasets}
