% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stackf}
\alias{stackf}
\title{Domain stack}
\format{
An object of class \code{ncdf4}
}
\usage{
stackf
}
\description{
NetCDF file (read by ncdf4) with points used to align CMAQ output to a grid when calculating emissions
}
\details{
To be used when calculating emissions from CMAQ output. The fields \code{COLPTc} and \code{ROWPTc} are used after emissions have been summed together and "flattened" from a 2D matrix to a 1D vector.
}
\keyword{datasets}
