% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{voc_mw}
\alias{voc_mw}
\title{Volatile Organic Compound (VOC) molecular weights}
\format{
A named vector
}
\usage{
voc_mw
}
\description{
A vector of commonly-used VOC acronyms and their corresponding molecular weights
}
\details{
Named vector of VOCs and their molecular weights.
}
\keyword{datasets}
